+++ src/RestMethods.php
@@ -13,16 +13,17 @@ declare(strict_types=1);
 
 namespace Dotclear\Plugin\relatedLinks;
 
+use Exception;
 class RestMethods
 {
     public static function addRelatedLink(array $get): array
     {
         if (empty($get['postId'])) {
-            throw new \Exception('No post ID');
+            throw new Exception('No post ID');
         }
 
         if (empty($get['linkId'])) {
-            throw new \Exception('No link ID');
+            throw new Exception('No link ID');
         }
 
         $manager = new RelatedLinks((int) $get['postId']);
@@ -34,11 +35,11 @@ class RestMethods
     public static function removeRelatedLink(array $get): array
     {
         if (empty($get['postId'])) {
-            throw new \Exception('No post ID');
+            throw new Exception('No post ID');
         }
 
         if (empty($get['linkId'])) {
-            throw new \Exception('No link ID');
+            throw new Exception('No link ID');
         }
 
         $manager = new RelatedLinks((int) $get['postId']);
@@ -50,7 +51,7 @@ class RestMethods
     public static function removeRelatedLinks(array $get): array
     {
         if (empty($get['postId'])) {
-            throw new \Exception('No post ID');
+            throw new Exception('No post ID');
         }
 
         $manager = new RelatedLinks((int) $get['postId']);
@@ -62,7 +63,7 @@ class RestMethods
     public static function getRelatedLinks(array $get): array
     {
         if (empty($get['postId'])) {
-            throw new \Exception('No post ID');
+            throw new Exception('No post ID');
         }
 
         $manager = new RelatedLinks((int) $get['postId']);